refactor: move directory watching into daemon core
